# docs/swagger.yaml
basePath: /
definitions:
...
  helpers.SwaggerError400Response:
    properties:
      message:
        example: Input tidak valid
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError401Response:
    properties:
      message:
        example: Token tidak ditemukan atau tidak valid
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError403Response:
    properties:
      code:
        example: ERR_PERMISSION_DENIED
        type: string
      message:
        example: 'Akses ditolak: Anda tidak memiliki akses untuk fitur ini'
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError404Response:
    properties:
      message:
        example: Data tidak ditemukan
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError500Response:
    properties:
      message:
        example: Terjadi kesalahan internal pada server
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerSuccessResponse:
    properties:
      data: {}
      message:
        example: Berhasil memproses data
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  model.Guru:
    properties:
      alamat:
        description: Deprecated Database Columns (Retained for Non-Destructive Migrations)
        example: Jl. Diponegoro No. 25
        type: string
      alamat_detail:
        example: Jl. Diponegoro No. 25
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      desa:
        example: Sidorejo Lor
        type: string
      gelar:
        example: S.Pd.
        type: string
      id:
        example: 1
        type: integer
      jenis_kelamin:
        example: Laki-laki
        type: string
      kabupaten:
        example: Salatiga
        type: string
      kecamatan:
        example: Sidorejo
        type: string
      mapel_ids:
        description: Virtual field for frontend multiselect checkboxes
        example:
        - 1
        - 2
        items:
          type: integer
        type: array
      nama:
        example: Ahmad Guru
        type: string
      nip:
        example: GR001
        type: string
      no_hp:
        example: "081234567890"
        type: string
      photo_url:
        example: /uploads/profile/guru_1.jpg
        type: string
      provinsi:
        description: Address Normalization Fields
        example: Jawa Tengah
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        example: 3
        type: integer
    type: object
...
  model.Kelas:
    properties:
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      jurusan:
        example: Rekayasa Perangkat Lunak
        type: string
      kapasitas:
        example: 36
        type: integer
      nama_kelas:
        example: X RPL 1
        type: string
      tingkat:
        example: X
        type: string
      total_siswa:
        example: 32
        type: integer
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      wali_kelas:
        $ref: '#/definitions/model.Guru'
      wali_kelas_id:
        example: 1
        type: integer
    type: object
...
  model.Siswa:
    properties:
      alamat:
        example: Jl. Diponegoro No. 25
        type: string
      alamat_detail:
        example: Jl. Diponegoro No. 25
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      desa:
        example: Sidorejo Lor
        type: string
      email:
        example: budi.santoso@sekolah.com
        type: string
      id:
        example: 1
        type: integer
      jenis_kelamin:
        example: Laki-laki
        type: string
      kabupaten:
        example: Salatiga
        type: string
      kecamatan:
        example: Sidorejo
        type: string
      kelas:
        $ref: '#/definitions/model.Kelas'
      kelas_id:
        example: 1
        type: integer
      nama:
        example: Budi Santoso
        type: string
      nis:
        example: "12345678"
        type: string
      no_hp:
        example: "081234567890"
        type: string
      photo_url:
        example: /uploads/profile/siswa_1.jpg
        type: string
      provinsi:
        description: Address Normalization Fields
        example: Jawa Tengah
        type: string
      tanggal_lahir:
        example: "2010-05-15"
        type: string
      tempat_lahir:
        description: Deprecated Database Columns (Retained for Non-Destructive Migrations)
        example: Salatiga
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      user_id:
        example: 2
        type: integer
    type: object
  model.SiswaResponseCreated:
    properties:
      data:
        $ref: '#/definitions/model.Siswa'
      message:
        example: Siswa berhasil ditambahkan
        type: string
      success:
        example: true
        type: boolean
      temporary_password:
        example: Siswa123!
        type: string
    type: object
  model.SiswaResponseList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Siswa'
        type: array
      message:
        example: Berhasil mengambil data siswa
        type: string
      success:
        example: true
        type: boolean
    type: object
  model.SiswaResponseSingle:
    properties:
      data:
        $ref: '#/definitions/model.Siswa'
      message:
        example: Berhasil mengambil detail siswa
        type: string
      success:
        example: true
        type: boolean
    type: object
  model.User:
    properties:
      alamat:
        example: Jl. Diponegoro No. 25, Salatiga
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      email:
        example: ahmad.guru@sekolah.com
        type: string
      gelar:
        example: S.Pd.
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      is_first_login:
        example: true
        type: boolean
      jenis_kelamin:
        example: Laki-laki
        type: string
      last_login_at:
        type: string
      nama:
        example: Ahmad Guru
        type: string
      nip:
        description: Ignored by GORM, populated for unified API login response
        example: GR001
        type: string
      nis:
        example: "12345678"
        type: string
      no_hp:
        example: "081234567890"
        type: string
      role:
        example: guru
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: praktikum@example.com
    name: Praktikum Pemrograman III
  description: Dokumentasi API backend Sistem Informasi Sekolah menggunakan Golang
    Fiber, GORM, PostgreSQL, dan JWT.
  title: API Sistem Informasi Sekolah
  version: "1.0"
paths:
...
  /api/siswa:
    get:
      consumes:
      - application/json
      description: Mengambil seluruh data siswa. Mendukung parameter pencarian dan
        filter kelas. Route ini dapat diakses secara publik, namun Guru/Wali Kelas
        akan mendapatkan data terbatas sesuai kelas masing-masing jika login.
      parameters:
      - description: Cari berdasarkan nama atau NIS
        in: query
        name: search
        type: string
      - description: Filter berdasarkan ID kelas
        in: query
        name: kelas_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SiswaResponseList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      summary: Ambil semua data siswa
      tags:
      - Siswa
    post:
      consumes:
      - application/json
      description: Menambahkan data siswa baru dan membuat user account dengan password
        acak sementara. Hanya dapat diakses oleh user yang terautentikasi (JWT).
      parameters:
      - description: Payload data siswa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Siswa'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.SiswaResponseCreated'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Tambah data siswa
      tags:
      - Siswa
  /api/siswa/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data siswa beserta user account terkait berdasarkan ID.
        Hanya dapat diakses oleh Admin (JWT + Admin role).
      parameters:
      - description: ID Siswa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.SwaggerSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.SwaggerError403Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Hapus data siswa
      tags:
      - Siswa
    get:
      consumes:
      - application/json
      description: Mengambil detail data satu siswa berdasarkan ID.
      parameters:
      - description: ID Siswa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SiswaResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      summary: Ambil data siswa berdasarkan ID
      tags:
      - Siswa
    put:
      consumes:
      - application/json
      description: Mengubah data siswa berdasarkan ID. Hanya dapat diakses oleh user
        yang terautentikasi (JWT).
      parameters:
      - description: ID Siswa
        in: path
        name: id
        required: true
        type: integer
      - description: Payload data siswa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Siswa'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SiswaResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ubah data siswa
      tags:
      - Siswa
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"